requirements.txt
python-telegram-bot[job-queue]==22.6
google-genai>=1.0.0
aiohttp>=3.9
aiosqlite>=0.19
python-dotenv>=1.0
httpx>=0.27
